Create group form when added to VO's embedded groups
* Moved Entry logic to registrar (needs to be handled in GUI)
* Empty forms was created when submitting application to embedded group
* Newely this happens already when group is allowed for auto registration
